+++ Shelby.tv/Library/GPUImage/GPUImageFilterPipeline.m
@@ -183,7 +183,7 @@
     id prevFilter = self.input;
     GPUImageFilter *theFilter = nil;
     
-    for (int i = 0; i < [self.filters count]; i++) {
+    for (NSUInteger i = 0; i < [self.filters count]; i++) {
         theFilter = [self.filters objectAtIndex:i];
         [prevFilter removeAllTargets];
         [prevFilter addTarget:theFilter];
